+++ sources/tuples/operations/magnitude.c
@@ -1,12 +1,24 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   magnitude.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kdaiane- < kdaiane-@student.42sp.org.br    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/26 20:27:38 by kdaiane-          #+#    #+#             */
+/*   Updated: 2023/09/26 20:27:39 by kdaiane-         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <minirt.h>
 
-float	magnitude(t_tuple v)
+double	magnitude(t_tuple v)
 {
-	float	x_squared;
-	float	y_squared;
-	float	z_squared;
-	float	w_squared;
-	float	result;
+	double	x_squared;
+	double	y_squared;
+	double	z_squared;
+	double	w_squared;
+	double	result;
 
 	x_squared = pow(v[0], 2);
 	y_squared = pow(v[1], 2);
